requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-celery-results==1.0.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
django-widget-tweaks==1.4.3
djutils==0.3.2
gunicorn==19.9.0
idna==2.8
kombu==4.2.2.post1
Pillow==5.4.1
psycopg2==2.7.6.1
python-bitcoinrpc==1.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
gnupg==2.3.0
django_countries==5.3.3
cryptowatch-client==0.1.4
django-star-ratings==0.7.0
pymysql==0.9.3
django-avatar